report-generator/app/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.22
python-docx>=1.2.0
lxml>=5.2.0
pydantic>=2.5
